Simple separation at order 1 testing
My code did not work for the simple separation.

Didier's did.

It looks like the issue is inside the X definition. When X is [-1, 1] (and other large ranges including both positive and negative numbers), the separation routine works and yields a 2nd order v function.

When X is [0,1] or [-0.3, 0.3], the separation routine does not work (primal infeasible).